requirements.txt
Flask
python-dotenv
openai-whisper
google-generativeai
torch
